refactor(git): split github.rs god-file into modules (#5891)
